🚚 rename files for consistency
