+++ app/node/store/src/main.ts
@@ -4,7 +4,7 @@ import git from "simple-git"
 import { Server as WebSocketServer } from "socket.io"
 
 import { serveFilestore } from "~/packages/socket-io.filestore/src/node/socket-filestore-node"
-import { serveSimpleGit } from "~/packages/socket-io.git/src/socket-io-git.node"
+import { serveSimpleGit } from "~/packages/socket-io.git/src/socket-git-node"
 
 import { logger } from "./logger"
 import { formatJson } from "./services/formatJson"
