+++ app/web/studio/src/services/git.ts
@@ -3,7 +3,7 @@ import { atom, useRecoilTransaction_UNSTABLE } from "recoil"
 import {
   isGitSocketError,
   initGitClientTools,
-} from "~/packages/socket-io.git/src/socket-io-git.web"
+} from "~/packages/socket-io.git/src/socket-git-recoil"
 import type { Transact } from "~/packages/hamr/src/recoil-tools/recoil-transaction-tools"
 
 import { socket } from "./socket"
