+++ packages/socket-io.git/node/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@socket-io.git/web",
-  "main": "../dist/git-io-node.js",
-  "types": "../dist/git-io-node.d.ts",
-  "module": "../dist/git-io-node.mjs"
+  "main": "../dist/socket-git-node.js",
+  "types": "../dist/socket-git-node.d.ts",
+  "module": "../dist/socket-git-node.mjs"
 }
