+++ packages/socket-io.git/package.json
@@ -10,7 +10,7 @@
     "README.md"
   ],
   "scripts": {
-    "build": "tsup --entry src/socket-io-git-node.ts --entry src/socket-io-git-web.ts --format cjs,esm --dts --external react,recoil,socket.io,socket.io-client",
+    "build": "tsup --entry src/socket-git-node.ts --entry src/socket-git-recoil.ts --format cjs,esm --dts --external react,recoil,socket.io,socket.io-client",
     "lint": "eslint",
     "test": "vitest",
     "test:once": "vitest run"
