+++ packages/socket-io.git/src/interface.ts
